+++ hello_imgui_cmake/overlay_vcpkg/hello-imgui/portfile.cmake
@@ -82,12 +82,13 @@ vcpkg_cmake_configure(
 
 vcpkg_cmake_install()
 
-vcpkg_cmake_config_fixup(CONFIG_PATH lib/cmake/hello_imgui PACKAGE_NAME "hello-imgui")  # should be active once himgui produces a config
+vcpkg_cmake_config_fixup(CONFIG_PATH lib/cmake/hello_imgui PACKAGE_NAME "hello_imgui")
 
 file(REMOVE_RECURSE
     "${CURRENT_PACKAGES_DIR}/debug/include"
     "${CURRENT_PACKAGES_DIR}/debug/share"
-    "${CURRENT_PACKAGES_DIR}/share/hello-imgui/hello_imgui_cmake/ios-cmake"
+    "${CURRENT_PACKAGES_DIR}/share/hello_imgui/hello_imgui_cmake/ios-cmake"
+    "${CURRENT_PACKAGES_DIR}/lib/cmake/hello-imgui"  # deprecated compat shim, not needed under vcpkg
 )
 
 vcpkg_install_copyright(FILE_LIST "${SOURCE_PATH}/LICENSE")